update
git-svn-id: http://svn.tudos.org/repos/oc/tudos/trunk@40 d050ee49-bd90-4346-b210-929a50b99cfc

+++ kernel/fiasco/src/jdb/jdb_utcb.cpp
@@ -51,7 +51,7 @@ Jdb_utcb::print(Thread *t)
       Vcpu_state *v = t->vcpu_state().kern();
       printf("\nVcpu-state-addr: %p\n", v);
       printf("state: %x    saved-state:  %x  sticky: %x\n",
-             v->state, v->_saved_state, v->sticky_flags);
+             (int)v->state, (int)v->_saved_state, (int)v->sticky_flags);
       printf("entry_sp = %lx    entry_ip = %lx  sp = %lx\n",
              v->_entry_sp, v->_entry_ip, v->_sp);
       v->_ts.dump();
